# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Capital Empreendedor",
    "description": "api de manipulação",
    "termsOfService": "http://localhost:3000/capital",
    "contact": {
      "email": "daniel_l0l@hotmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/users",
      "description": "Api de test"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "summary": "Cadastro de cliente",
        "description": "cadastra um novo cliente",
        "tags": ["users"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              },
              "examples": {
                "user": { 
                  "value": {
                    "name": "Cliente 1",
                    "email": "cliente1@example.com",
                    "isActive": true,
                    "phone": "(19) 99999-9999",
                    "revenue": 1000000,
                    "agreedTerms": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "User Already exists!"
          },
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/user"
                }  
              }
            }
          }
        }
      }
    },
    "/users/findAll": {
      "get": {
        "description": "listar cliente",
        "summary": "listar cliente",
        "tags": ["users"],
        "parameters": [
         
        ],
        "responses": {
          "200": {
          "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "arrey",
                  "item": {
                    "$ref": "#/components/schemas/user"      
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Busca de um cliente pelo id",
        "summary": "Busca de um cliente pelo id",
        "tags": ["users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Nome do cliente pelo id",
            "required": true
          }
        ],
        "responses": {
          "200": {
          "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/user"      
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Alterar o cliente pelo id",
        "summary": "Alterar um cliente pelo id",
        "tags": ["users"],
        
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Nome do cliente para pelo id",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              },
              "examples": {
                "user": { 
                  "value": {
                    "name": "Cliente 1",
                    "email": "cliente1@example.com",
                    "isActive": true,
                    "phone": "(19) 99999-9999",
                    "revenue": 1000000,
                    "agreedTerms": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok",  
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/user"         
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Excluir um cliente pelo id",
        "summary": "Excluir um cliente pelo id",
        "tags": ["users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Delete cliente pelo id",
            "required": true
          }
        ],
        "responses": {
          "200": {
          "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/user"      
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "user": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "phone": {
            "type": "string"
          },
          "revenue": {
            "type": "number"
          },
          "agreedTerms": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          }
        }
      } 
    }
  }
}